Namespace ButtonSpecHelper to Spec::Support::ButtonHelper
