--- spec/support/button_spec_helper.rb
@@ -1,8 +1,0 @@
-module ButtonSpecHelper
-  def setup_view_context_with_sandbox(sandbox)
-    view_context = double('ApplicationHelper')
-    view_context.class.send(:include, ApplicationHelper)
-    view_context.instance_eval { @sb = sandbox }
-    view_context
-  end
-end
